Make die work with nounset
I like to keep `set -euo pipefail` at the top of my scripts. This causes
a couple of errors with argbash's die function reading unset variables:

1. it sets a default for the exit code like in a way that
   unconditionally reads the second argument

    local _ret=$2  # <---- can fail if only one arg passed in
    test -n "$_ret" || _ret=1

2. it reads unconditionally from _PRINT_HELP, which may not be set

This commit changes both of these to use default values in parameter
expansion.  This construct is part of POSIX, and so should work in both
bash and posix mode.

Additionally fix a shellcheck quote warning.

Tested: Verified the generated die function works with nounset.