Add flake.nix file for Nix/NixOS users
This flake defines a development shell for developing argbash. It
ensures that all dependencies used to build this project are pinned to a
specific version (via a lock file). In this case, make, m4, and autoconf,
and sphinx.

This guarantees deterministic build behavior; builds always use the same
exact dependencies until the lock file is updated. Any updates to the
dependencies on a system level will not be able to break our build if an
API or a behavior of such a dependency has changed.

The nixpkgs repository already has a build script for argbash here:

https://github.com/NixOS/nixpkgs/tree/master/pkgs/by-name/ar/argbash

But that is just for building the tool so it can be used. A devshell can
be defined and used from there, but this patch allows it to be updated when
new dependencies come in, allowing us to keep the flake in sync with
development of this tool.